requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
firebase-admin==6.6.0
pydantic==2.9.2
python-dotenv==1.0.1